# Validate members once, in `NewMember`