This is an alternative attempt to #2505 at at improving the number of calls to `validate()` in `baggage.Member`, calling it only when the member is added to the baggage.
Right now, we call it twice. Once within `baggage.NewMember`, and another time within `baggage.New`.

Closes #2498