@dmathieu 2022-01-20
I have added an `isEmpty`, which is used to check whether the provided member and property are empty or not, so they don't need revalidation but also can't be added to their parent struct if they are empty.