@MadVikingGod 2022-02-02
So as it stands, this change has a net negative effect on parsing.  For this to proceed, you will have to do a bit of profiling and improve what is already here.

```
name         old time/op    new time/op    delta
New-4          3.15µs ± 5%    3.12µs ± 3%     ~     (p=0.489 n=9+9)
NewMember-4     363ns ± 5%     360ns ± 3%     ~     (p=0.605 n=9+9)
Parse-4        6.29µs ± 6%    7.40µs ± 4%  +17.61%  (p=0.000 n=10+9)

name         old alloc/op   new alloc/op   delta
New-4            832B ± 0%      832B ± 0%     ~     (all equal)
NewMember-4     0.00B          0.00B          ~     (all equal)
Parse-4        1.35kB ± 0%    1.43kB ± 0%   +5.94%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
New-4            16.0 ± 0%      16.0 ± 0%     ~     (all equal)
NewMember-4      0.00           0.00          ~     (all equal)
Parse-4          19.0 ± 0%      20.0 ± 0%   +5.26%  (p=0.000 n=10+10)
```