@dmathieu 2022-02-03
So, I've done two changes:

* `New` actually needs to very `hasData` and not call `validate()`. If we call validate, with `hasData` being true, we rerun the validation (hence no performance improvement in your benchstat).
* We need to check the regex themselves and not use `NewMember` in `parseMember`. Doing the latter reruns the validation, both on the member and on the properties, hence the performance loss.

Here is my benchstat with the latest changes:


```
name          old time/op    new time/op    delta
New-12          2.45µs ± 1%    1.22µs ± 2%  -50.29%  (p=0.000 n=10+10)
NewMember-12     288ns ± 1%     290ns ± 2%     ~     (p=0.315 n=9+10)
Parse-12        4.92µs ± 1%    4.94µs ± 1%     ~     (p=0.589 n=10+9)

name          old alloc/op   new alloc/op   delta
New-12            832B ± 0%      832B ± 0%     ~     (all equal)
NewMember-12     0.00B          0.00B          ~     (all equal)
Parse-12        1.36kB ± 0%    1.36kB ± 0%     ~     (p=0.875 n=10+10)

name          old allocs/op  new allocs/op  delta
New-12            16.0 ± 0%      16.0 ± 0%     ~     (all equal)
NewMember-12      0.00           0.00          ~     (all equal)
Parse-12          19.0 ± 0%      19.0 ± 0%     ~     (all equal)
```